containers/base/requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
pyyaml==6.0.1
